Refactor GPTQ Quantizer, remove lm_eval
Summary:

refactor GPTQ code, remove lm_eval dependency of gptq, remove model
dependency of InputRecorder made GPTQ work with gpt-fast. also fixed
model so its kv_cache doesn't break gptq.

Test Plan:
python test/quantization/test_quant_api.py

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: d6eb81037e37ba556bd37a3b248c51482be7c68d
Pull Request resolved: https://github.com/pytorch-labs/ao/pull/103
